mixed continuous crawler
